keystore: allow system to clear keys even while encrypted
Since we can store keys that are unencrypted, we should allow the
clearing of data for apps when the keystore is still in a LOCKED state.
Also allow the system user to delete keys since this is necessary for
system maintenance when application data is cleared or the application
is uninstalled.

(cherry picked from commit e17c25459fc0f200134e10a1aaef12fa8f930f04)

Bug: 15751553
Change-Id: Id02bc9992bd529e79be7a09d2bce208942d42b84