Update the IKeystoreService Binder API
The Framework side of IKeystoreService is moving to using an aidl to
generate the stubs instead of hand written code, as a side effect some
changes need to be made to keep Keystore consistent.

The interface name is to change to android.security.IKeystoreService

The deserialization of arguments for generate is changed to be based off
a Parcelable object instead of a byte[][] since aidl/Parcel don't
support writing byte[][] directly.

Change-Id: I3fd59f2e4b9257d194ca9c36e9597d8d5144aada