keystore: log pid and uid when auditing SE Linux denial
When debugging SE Linux rules, the pid and uid of the source context are
helpful in determining what was involved in generating the policy violation.
This information was absent from the keystore logs.

To remedy this, we add pid and uid to the logs:

avc:  denied  { exist } for pid=571 uid=1000 ...

Change-Id: Ic22128720aa8ac225f26896f9e710783c8ab4f70
Signed-off-by: William Roberts <william.c.roberts@intel.com>