Add attestation application id for key attestation
This patch adds functionality for gathering an application id
for the attestation of a key that is bound to an application
in the keystore.

Keystore gathers the information package name, package version,
and signing certificates of the calling app from the package manager.
It then DER encodes the information and appends it to attestation
parameters.

Bug: 22914603
Change-Id: I9fe1d8f97ee1dfa79284bcf751f86631c94d4174