Correct missed return code path.
ag/2835967 corrected a lost authorization check result introduced during
a refactor, but in the process failed to return the auth check result in
one codepath, causing the Java layer above to fail to throw the
exception.  That in turn broke work profile password removal.  This CL
sets the return code for cases where the authorization check failed.

Bug: 65348783
Test: Manually tested with TestDPC, add and remove passcode
Change-Id: I846b154c8cbcd9a73cd12b9a4376616dacf62fbb