Fix misnomer and resulting bug in keystore.cpp
On the upgrade path from keystore blobs version 0, the userId was
interpreted as uid which resulted in it being converted to a userId a
second time. This would result in keys belonging to other users
being assigned to the main user.

This patch fixes the bug and the misnomer that lead to the confusion.

Change-Id: I91975310b6140929dcb6820aa8bd4c28b8e5de5e