Allow Device IDs in Key attestation request
Change Keystore to allow inclusion of device identifiers in the key
attestation request - but if, and only if, the request is coming from
the system context.

This was not supported previously as Device ID attestation was done
separately with an ephemeral key.
But as KeyChain-generated keys need to include device identifiers,
Keystore is changed to allow that, but only for attestation requests
from the system context.

Bug: 63388672
Test: cts-tradefed run commandAndExit cts-dev -a armeabi-v7a -m CtsDevicePolicyManagerTestCases -t com.android.cts.devicepolicy.DeviceOwnerTest#testKeyManagement
Change-Id: Ibbcf0c51931b644b98ab78b21e92a1edf571dc05