Fix strict keymaster configuration checking.
With keymaster 4 a new security level strongbox was introduced.
When strongbox is present a TEE Keymaster implementation must
also be present. Checking this rule strictly poses complications for the
teams integrating these implementations.

This patch relaxes the Keystores bootstrapping such that Keymaster4
strongbox implementations are ignored as long as there is no Keymaster4
TEE implementation present. This allows Keystore to fall back to
Keymaster3 even though Keymaster4 strongbox is present.

Change-Id: I06215cb4adfc09e90aea0c24866294a91c5601bc