Fix KeyStoreClientImpl::doesKeyExist to return correct result
The KeyStoreService returns NO_ERROR if the key was found,
and another response code otherwise. All of these are
mapped to non-zero values. As a result, if a key's
existence was queried it would always respond "true",
regardless if it exists or not or if there was a permissions
error.

Test: Key existence can be successfully checked with the
      keystore_cli_v2 tool.

Change-Id: Iffc2e155a61354f1fbffbece093b19e5cbc537fd